pyproject.toml
[project]
name = "llm-code-review"
version = "0.1.0"
description = ""
authors = [
    {name = "Marcus",email = "help@google.com"}
]
requires-python = ">=3.14"
dependencies = [
    "python-dotenv (>=1.2.2,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
